一些优化
1.解决了数据库字符集的问题.
2.增加了角色换武器功能.

+++ Server/Src/LogicServer/GlobalDataMgr.cpp
@@ -32,7 +32,7 @@ BOOL CGlobalDataManager::LoadGlobalData(CppMySQL3DB& tDBConnection)
 	if(!QueryResult.eof())
 	{
 		dwMaxGuid = QueryResult.getInt64Field("maxguid");
-		dwMaxOnline = QueryResult.getInt64Field("maxonline");
+		dwMaxOnline = QueryResult.getIntField("maxonline");
 	}
 
 	if(dwMaxGuid == 0)
